allow no-props in the stream-insert-callback
The callback needs to be used in order to advance state.  When inserting tool
call results into the buffer (which will persist their turns), they do not need
to be propertized like a response.

This exposes the fact that the default callback is too generic and we probably
need one per-frontend.  The reason we use this callback is because we're not
sure where to insert tool results (and only two formats are even supported) but
honestly the callback doesn't know either because it's not front-end aware.