remove chat persistence feature
This might be back later, but it will definitely not be the same form.  Org
files would be a good way to encode multiple prompts and tools, using their
metadata and structure to create sub-types of operations.

But the existing save / restore adds messy data to context and is not meant to
be used in the way described above.